+++ scripts/winesapos-install.sh
@@ -853,7 +853,7 @@ elif [[ "${WINESAPOS_DE}" == "hyprland" ]]; then
 
 elif [[ "${WINESAPOS_DE}" == "plasma" ]]; then
     echo "Installing the KDE Plasma desktop environment..."
-    pacman_install_chroot plasma-meta plasma-nm
+    pacman_install_chroot plasma-meta plasma-nm maliit-keyboard
     # Dolphin file manager and related plugins.
     pacman_install_chroot dolphin ffmpegthumbs kdegraphics-thumbnailers konsole
     chroot "${WINESAPOS_INSTALL_DIR}" crudini --ini-options=nospace --set /etc/xdg/konsolerc "Desktop Entry" DefaultProfile Vapor.profile
@@ -904,7 +904,14 @@ elif [[ "${WINESAPOS_DE}" == "plasma" ]]; then
     done
     chroot "${WINESAPOS_INSTALL_DIR}" groupadd nopasswdlogin
     chroot "${WINESAPOS_INSTALL_DIR}" usermod -a -G nopasswdlogin "${WINESAPOS_USER_NAME}"
-    echo "InputMethod=qtvirtualkeyboard" | sudo tee "${WINESAPOS_INSTALL_DIR}"/etc/sddm.conf.d/winesapos.conf
+    "${CMD_PACMAN_INSTALL[@]}" maliit-keyboard
+    echo "[General]
+DisplayServer=wayland
+GreeterEnvironment=QT_WAYLAND_SHELL_INTEGRATION=layer-shell
+
+[Wayland]
+CompositorCommand=kwin_wayland --drm --no-lockscreen --no-global-shortcuts --locale1 --inputmethod maliit-keyboard" > "${WINESAPOS_INSTALL_DIR}"/etc/sddm.conf.d/winesapos.conf
+    echo "KWIN_IM_SHOW_ALWAYS=1" >> "${WINESAPOS_INSTALL_DIR}"/etc/environment
     echo "Configuring passwordless login complete."
 
     echo "Installing the KDE Plasma desktop environment complete."